Use `ureq` instead of `reqwest::blocking` for oauth2 http requests
`reqwest::blocking` will make the app panic if called from an async environment. Since git operations must use a synchronous auth agent, using `ureq` is probably our best option here.
