+++ crates/moss-git-hosting-provider/Cargo.toml
@@ -13,8 +13,8 @@ git2.workspace = true
 url.workspace = true
 dotenv.workspace = true
 serde_json.workspace = true
-oauth2 = { workspace = true, features = ["reqwest-blocking"] }         # Using blocking for easy testing here
-reqwest = { workspace = true, features = ["blocking", "json"] }
+oauth2 = { workspace = true, features = ["ureq"] }
+reqwest = { workspace = true, features = ["json"] }
 tokio = { workspace = true, features = ["rt-multi-thread", "macros"] }
 serde = { workspace = true, features = ["derive"] }
 webbrowser.workspace = true
