+++ view/desktop/bin/src/plugins.rs
@@ -24,6 +24,9 @@ pub mod plugin_log {
             .level_for("mio", log::LevelFilter::Off)
             .level_for("reqwest", log::LevelFilter::Off)
             .level_for("keyring", log::LevelFilter::Off)
+            .level_for("ureq", log::LevelFilter::Off)
+            .level_for("rustls", log::LevelFilter::Off)
+            .level_for("webbrowser", log::LevelFilter::Off)
             .with_colors(ColoredLevelConfig::default())
             .level(if is_dev() {
                 log::LevelFilter::Trace
